WIP: split the tests in mathematical_program_test.cc
