+++ drake/solvers/test/mathematical_program_test_util.h
@@ -15,6 +15,15 @@
 namespace drake {
 namespace solvers {
 namespace test {
+// TODO(#2274) Fix NOLINTNEXTLINE(runtime/references).
+void CheckSolverType(MathematicalProgram& prog,
+                     std::string desired_solver_name) {
+  std::string solver_name;
+  int solver_result;
+  prog.GetSolverResult(&solver_name, &solver_result);
+  EXPECT_EQ(solver_name, desired_solver_name);
+}
+
 void RunSolver(MathematicalProgram* prog,
                const MathematicalProgramSolverInterface& solver) {
   if (solver.available()) {
